Expose the Schema and Dynamic APIs in the PCL
