+++ Realm.Shared/Schema/Property.cs
@@ -16,9 +16,7 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-using System;
 using System.Diagnostics;
-using System.Diagnostics.Contracts;
 using System.Reflection;
 
 namespace Realms.Schema
