+++ Realm.Shared/Schema/PropertyType.cs
@@ -16,10 +16,6 @@
 //
 ////////////////////////////////////////////////////////////////////////////
 
-using System;
-using System.Diagnostics.Contracts;
-using System.Linq;
-
 namespace Realms.Schema
 {
     public enum PropertyType : byte
@@ -34,69 +30,6 @@ namespace Realms.Schema
         Date = 8,
         Object = 12,
         Array = 13
-    }
-
-    internal static class PropertyTypeEx
-    {
-        public static PropertyType ToPropertyType(this Type type, out bool isNullable, out Type innerType)
-        {
-            if (type == null) throw new ArgumentNullException(nameof(type));
-            Contract.EndContractBlock();
-
-            innerType = null;
-
-            var nullableType = Nullable.GetUnderlyingType(type);
-            isNullable = nullableType != null;
-            if (isNullable)
-                type = nullableType;
-
-            switch (Type.GetTypeCode(type))
-            {
-                case TypeCode.Boolean:
-                    return PropertyType.Bool;
-                case TypeCode.Char:
-                case TypeCode.Byte:
-                case TypeCode.Int16:
-                case TypeCode.Int32:
-                case TypeCode.Int64:
-                    return PropertyType.Int;
-                case TypeCode.Single:
-                    return PropertyType.Float;
-                case TypeCode.Double:
-                    return PropertyType.Double;
-                case TypeCode.String:
-                    isNullable = true;
-                    return PropertyType.String;
-            }
-
-            if (type == typeof(DateTimeOffset))
-                return PropertyType.Date;
-
-            if (type == typeof(byte[]))
-            {
-                isNullable = true;
-                return PropertyType.Data;
-            }
-
-            if (type.BaseType == typeof(RealmObject))
-            {
-                isNullable = true;
-                innerType = type;
-                return PropertyType.Object;
-            }
-
-            if (type.IsGenericType)
-            {
-                var definition = type.GetGenericTypeDefinition();
-                if (definition == typeof(RealmList<>))
-                {
-                    innerType = type.GetGenericArguments().Single();
-                    return PropertyType.Array;
-                }
-            }
-
-            throw new ArgumentException($"The property type {type.Name} cannot be expressed as a Realm schema type", nameof(type));
-        }
     }
 }
